ai-agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
openai>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
